requirements.txt
langchain==0.0.178
loguru==0.7.0
pandas==1.5.3
protobuf==3.20.3
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.22.0
chromadb